+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/effects/offense/crit/CriticalDamageEffect.java
@@ -27,7 +27,7 @@ public class CriticalDamageEffect extends BaseDamageEffect {
 
     @Override
     public DamageEffect activate(DamageEffect effect, StatData data, Stat stat) {
-        effect.percentIncrease += data.getAverageValue();
+        effect.increaseByPercent(data.getAverageValue());
         return effect;
     }