+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/resources/PlusResourceOnKill.java
@@ -74,7 +74,7 @@ public class PlusResourceOnKill extends Stat {
 
         @Override
         public DamageEffect activate(DamageEffect effect, StatData data, Stat stat) {
-            effect.addToRestore(new ConditionalRestoreResource(resource, data.getAverageValue(), x -> x.target.isDead()));
+            effect.addToRestore(new ConditionalRestoreResource(ConditionalRestoreResource.RestoreType.LEECH, resource, data.getAverageValue(), x -> x.target.isDead()));
             return effect;
         }