+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/resources/ResourceLeech.java
@@ -84,7 +84,7 @@ public class ResourceLeech extends Stat implements IGenerated<Stat> {
         @Override
         public DamageEffect activate(DamageEffect effect, StatData data, Stat stat) {
             float amount = data.getAverageValue() * effect.number / 100F;
-            effect.addToRestore(new ConditionalRestoreResource(info.resource, amount));
+            effect.addToRestore(new ConditionalRestoreResource(ConditionalRestoreResource.RestoreType.LEECH, info.resource, amount));
             return effect;
         }
 
@@ -100,7 +100,7 @@ public class ResourceLeech extends Stat implements IGenerated<Stat> {
 
         @Override
         public int GetPriority() {
-            return Priority.Last.priority;
+            return Priority.AlmostLast.priority;
         }
     }